server/index.ts => {server/index.ts, server/server.ts, server/factory.ts}
